PRO-10170 fix: error with rendering tests with initial `runHash`
fix: catch errors from original `actualValue` in `expect`
